Restructure cos_lock library
Changed cos_lock to cos_sync in attempt to make the library as generic as possible
for future use. There are still some functions, specifically ones related to managing
the intrrupt bitmask that need to be moved to their own file to make cos_sync truely
generic.

cos_sync also went through a second draft. Much of the code has been reworked in an
attempt to tackle concurency issues.