Share debug information and update cos type defs
This patch is not meant to be a fix for our syconization problems!
This patch is meant to push the current debugging state so that others
can assist in the debugging process.

This patch also includes a location change from the capid_t typedefs.
The move was motivated by needing to access some capid_t's without wanting
to include all of the cos_kernel_api. This move makes sense in a design
perspective as type defs should be logically located within cos_types.h which
cos_kernel_api.h already includes so no further change needs to be made to
cos_kernel_api.h